# Add table/instance filtering to speed up debugging process.