This makes it so you can do something like:

```fish
> xbrl_extract $PUDL_INPUT/2021/ferc1-xbrl-2021.zip foo.sqlite --instance-pattern "PJM" --requested-tables "electric_operations_and_maintenance_expenses_320_duration" "identification_001_duration" --metadata-path foo-md.json
```

And take a look-see at what the extractor is putting out, as well as stuff breakpoints in places without having to add conditional breakpoints like "if the table is blah and the instance is blah, then breakpoint()".


Some timings just for fun:

```
> time xbrl_extract $PUDL_INPUT/2021/ferc1-xbrl-2021.zip foo.sqlite --instance-pattern "PJM" --requested-tables "electric_operations_and_maintenance_expenses_320_duration" "identification_001_duration" --metadata-path foo-md.json
2023-09-27 13:01:52 [    INFO] catalystcoop.ferc_xbrl_extractor.xbrl:265 Parsing taxonomy from https://eCollection.ferc.gov/taxonomy/form1/2022-01-01/form/form1/form-1_2022-01-01.xsd
2023-09-27 13:01:56 [    INFO] catalystcoop.ferc_xbrl_extractor.xbrl:149 Finished batch 1/1

________________________________________________________
Executed in    5.89 secs    fish           external
   usr time    7.20 secs    0.31 millis    7.20 secs
   sys time    2.07 secs    1.75 millis    2.07 secs

```