@jdangerx 2023-09-27
Sort of seems like I probably broke some of this behavior that used to exist, since there was some old logic to hook into. But I guess it's fixed now! I'm gonna add a test to make codecov happy and then merge.